Make sure that the web browser doesn't write to stdout
Some browsers write to stdout, which is used  by git-annex for the
special remote protocol. So, when it happens, it causes git-annex to
stop because of some "invalid commands" in the special remote.

A simple yet effective solution is to start the webbrowser from an
entirely different process which has its stdout (and stderr, just to
avoid polluting our output) redirected to /dev/null.

Fixes #3.